refactor(sql): move connection pool and open_connection into sql_pool
Move the connection pool implementation (POOL_MAX_IDLE, POOL_MAX_IDLE_SECS,
_pool, _pool_lock, _pool_enabled, _pool_time, _is_alive, reset_pool,
_pool_checkout, _pool_checkin, _PooledConnection, _make_pool_key) and
open_connection from sql.py into sql_pool.py.

sql.py becomes a thin runner layer (SqlTarget, _with_connect_retry, run_query,
run_statements) with re-export shims so all public and test-visible names keep
resolving through fabric_dw.sql.*.

Retargeted patch sites (same foot-gun class as PR C — bare-name calls resolve
against the module where the callee is defined):
- fabric_dw.sql.get_sql_token_struct (7 sites in TestOpenConnectionOidcTokenInjection
  and TestOpenConnectionTimeoutAPI) -> fabric_dw.sql_pool.get_sql_token_struct
- fabric_dw.sql._pool_time (2 sites: TestPoolBehaviourIdleEviction,
  TestPoolCheckinEvictsStale) -> fabric_dw.sql_pool._pool_time
- fabric_dw.sql.POOL_MAX_IDLE set (2 sites: TestPoolBehaviourMaxIdle try/finally)
  -> fabric_dw.sql_pool.POOL_MAX_IDLE

Patches that remain on fabric_dw.sql.* and still work:
- fabric_dw.sql.open_connection — _with_connect_retry (in sql.py) calls
  open_connection by bare name, resolving sql.py's shim binding; patching it
  intercepts the call correctly.
- fabric_dw.sql._pool_enabled() calls — tests that call the function via
  _sql_module._pool_enabled() get the same function object; correct result.
- from fabric_dw.sql import _pool, _pool_lock — shim gives same object reference.

Part of #828
